refactor: commands/field_refs.rs を責務別に 7 ファイルへ分割 (#16)
1,397 行の field_refs.rs を以下の 7 ファイルへ分割。
ロジック変更なし・全テスト (363 件) パス確認済み。

- mod.rs             : 公開型 7 個 + pub use によるコマンド再公開
- helpers.rs         : 共有ヘルパー (fetch_occ_names, field_ref_matches 等)
- script_and_calc.rs : get_field_refs, get_field_calc_refs
- layout.rs          : get_field_layout_refs, resolve_layout_field
- relationships.rs   : get_field_relationship_keys
- unused_fields.rs   : list_unused_fields（最大・最複雑ロジック）
- debug.rs           : get_layout_ref_debug_info

Tauri の generate_handler! が __cmd__* を参照するため、
mod.rs で __cmd__* も明示的に pub use している。
