--- src-tauri/src/commands/field_refs.rs
@@ -1,1397 +1,0 @@
-﻿//! フィールド参照解析コマンド。
-
-use rusqlite::{params, OptionalExtension as _};
-use serde::{Deserialize, Serialize};
-
-use crate::{commands::CommandError, AppState};
-
-// ---------------------------------------------------------------------------
-// 公開型
-// ---------------------------------------------------------------------------
-
-/// フィールドがリレーションキーとして使用されているリレーションの情報。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FieldRelKeyRef {
-    pub relationship_id: i64,
-    pub relationship_name: String,
-    pub left_table: String,
-    pub right_table: String,
-    pub operator: String,
-    /// "left" or "right" — このフィールドがキーとして使われている側
-    pub side: String,
-}
-
-/// 未使用フィールドの情報。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UnusedFieldRow {
-    pub table_name: String,
-    pub field_name: String,
-    pub field_type: String,
-    pub data_type: String,
-    pub field_id: i64,
-}
-
-/// このフィールドを計算式で参照している他フィールドの情報。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FieldCalcRef {
-    pub field_id: i64,
-    pub field_name: String,
-    pub table_name: String,
-    pub table_id: i64,
-}
-
-/// フィールドを参照しているスクリプトの一覧。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FieldRefScript {
-    pub script_id: i64,
-    pub script_name: String,
-}
-
-/// フィールドのテーブルを使用しているレイアウトの一覧。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FieldRefLayout {
-    pub layout_id: i64,
-    pub layout_name: String,
-}
-
-/// テーブルオカレンス名とフィールド名から解決したフィールドの位置情報。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FieldLocation {
-    pub table_id: i64,
-    pub field_id: i64,
-    pub table_name: String,
-}
-
-/// レイアウトフィールド参照のデバッグ情報。
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LayoutRefDebugInfo {
-    pub occurrence_count: i64,
-    pub layout_field_ref_count: i64,
-    pub sample_occurrences: Vec<String>,
-    pub sample_field_refs: Vec<String>,
-}
-
-// ---------------------------------------------------------------------------
-// Tauri コマンド
-// ---------------------------------------------------------------------------
-
-/// フィールドを参照しているスクリプトの一覧を返す。
-///
-/// オカレンス名経由の検索 + 識別子境界チェックにより誤検知・漏れを防ぐ。
-#[tauri::command]
-pub async fn get_field_refs(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-    table_name: String,
-    field_name: String,
-) -> Result<Vec<FieldRefScript>, CommandError> {
-    if table_name.is_empty() || field_name.is_empty() {
-        return Ok(vec![]);
-    }
-    let db = super::lock_db(&state)?;
-
-    // 1. ベーステーブルに紐づく全オカレンス名を取得
-    let occ_names =
-        fetch_occ_names(&db.conn, project_id, &table_name).map_err(CommandError::from)?;
-
-    // 2. 全スクリプトステップを取得
-    let mut stmt = db
-        .conn
-        .prepare(
-            "SELECT s.id, s.name,
-                    COALESCE(ss.step_text, ''),
-                    COALESCE(ss.calculation, '')
-             FROM scripts s
-             JOIN script_steps ss ON ss.script_id = s.id
-             WHERE s.project_id = ?1",
-        )
-        .map_err(CommandError::from)?;
-    let step_rows: Vec<(i64, String, String, String)> = stmt
-        .query_map(params![project_id], |row| {
-            Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?))
-        })
-        .map_err(CommandError::from)?
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(CommandError::from)?;
-
-    // 3. Rust でフィルタリング（DISTINCT script_id）
-    let mut seen = std::collections::HashSet::new();
-    let results = step_rows
-        .into_iter()
-        .filter(|(_, _, step_text, calc)| {
-            field_ref_matches(step_text, &occ_names, &field_name)
-                || field_ref_matches(calc, &occ_names, &field_name)
-        })
-        .filter_map(|(script_id, script_name, _, _)| {
-            seen.insert(script_id).then_some(FieldRefScript {
-                script_id,
-                script_name,
-            })
-        })
-        .collect::<Vec<_>>();
-
-    let mut results = results;
-    results.sort_by(|a, b| a.script_name.cmp(&b.script_name));
-    Ok(results)
-}
-
-/// このフィールドを計算式（calculation）で参照している他フィールドの一覧を返す。
-///
-/// オカレンス名経由の検索 + 識別子境界チェックにより誤検知・漏れを防ぐ。
-/// 自分自身は除外する。
-#[tauri::command]
-pub async fn get_field_calc_refs(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-    table_name: String,
-    field_name: String,
-) -> Result<Vec<FieldCalcRef>, CommandError> {
-    if table_name.is_empty() || field_name.is_empty() {
-        return Ok(vec![]);
-    }
-    let db = super::lock_db(&state)?;
-
-    // 1. ベーステーブルに紐づく全オカレンス名を取得
-    let occ_names =
-        fetch_occ_names(&db.conn, project_id, &table_name).map_err(CommandError::from)?;
-
-    // 2. 全フィールド（calculation 非空・自分自身除く）を取得
-    let mut stmt = db
-        .conn
-        .prepare(
-            "SELECT f.id, f.name, bt.name, bt.id, f.calculation
-             FROM fields f
-             JOIN base_tables bt ON bt.id = f.table_id
-             WHERE bt.project_id = ?1
-               AND f.calculation IS NOT NULL
-               AND f.calculation != ''
-               AND NOT (bt.name = ?2 AND f.name = ?3)
-             ORDER BY bt.name, f.name",
-        )
-        .map_err(CommandError::from)?;
-    let candidates: Vec<(i64, String, String, i64, String)> = stmt
-        .query_map(params![project_id, table_name, field_name], |row| {
-            Ok((
-                row.get(0)?,
-                row.get(1)?,
-                row.get(2)?,
-                row.get(3)?,
-                row.get(4)?,
-            ))
-        })
-        .map_err(CommandError::from)?
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(CommandError::from)?;
-
-    // 3. Rust でフィルタリング
-    let results = candidates
-        .into_iter()
-        .filter(|(_, _, _, _, calc)| field_ref_matches(calc, &occ_names, &field_name))
-        .map(
-            |(field_id, field_name, table_name, table_id, _)| FieldCalcRef {
-                field_id,
-                field_name,
-                table_name,
-                table_id,
-            },
-        )
-        .collect();
-    Ok(results)
-}
-
-/// フィールドが配置されているレイアウトの一覧を返す。
-///
-/// 検索手順：
-/// 1. `table_name`（ベーステーブル名）に対応するオカレンス名一覧を取得
-/// 2. メインテーブル（`table_occurrence_name`）がそのオカレンスの1つであるレイアウトを絞り込む
-/// 3. さらに、そのレイアウト上に `field_name` が配置されているかを確認
-#[tauri::command]
-pub async fn get_field_layout_refs(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-    table_name: String,
-    field_name: String,
-) -> Result<Vec<FieldRefLayout>, CommandError> {
-    if table_name.is_empty() || field_name.is_empty() {
-        return Ok(vec![]);
-    }
-    let db = super::lock_db(&state)?;
-    get_field_layout_refs_inner(&db.conn, project_id, &table_name, &field_name)
-        .map_err(CommandError::from)
-}
-
-/// テーブルオカレンス名とフィールド名からフィールドの DB ID・テーブル DB ID・ベーステーブル名を解決する。
-#[tauri::command]
-pub async fn resolve_layout_field(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-    occurrence_name: String,
-    field_name: String,
-) -> Result<Option<FieldLocation>, CommandError> {
-    let db = super::lock_db(&state)?;
-    resolve_layout_field_inner(&db.conn, project_id, &occurrence_name, &field_name)
-        .map_err(CommandError::from)
-}
-
-/// フィールドがリレーションキーとして使用されているリレーション一覧を返す。
-///
-/// `table_name` はベーステーブル名。
-/// `join_predicates` を `relationships` / `table_occurrences` と結合して検索する。
-#[tauri::command]
-pub async fn get_field_relationship_keys(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-    table_name: String,
-    field_name: String,
-) -> Result<Vec<FieldRelKeyRef>, CommandError> {
-    if table_name.is_empty() || field_name.is_empty() {
-        return Ok(vec![]);
-    }
-    let db = super::lock_db(&state)?;
-    get_field_relationship_keys_inner(&db.conn, project_id, &table_name, &field_name)
-        .map_err(CommandError::from)
-}
-
-/// レイアウト・リレーション・スクリプト計算式・フィールド計算式・バリューリストから
-/// 一度も参照されていないフィールドの一覧を返す。
-///
-/// 検査対象:
-/// - layout_field_refs / layout_objects（レイアウト配置）
-/// - join_predicates（リレーション結合キー）
-/// - value_list_field_refs（バリューリストのフィールドソース）
-/// - script_steps.calculation / step_text, fields.calculation / auto_enter_calc / val_calc
-///   （OccName::FieldName パターンのみ検出）
-#[tauri::command]
-pub async fn list_unused_fields(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-) -> Result<Vec<UnusedFieldRow>, CommandError> {
-    let db = super::lock_db(&state)?;
-    list_unused_fields_inner(&db.conn, project_id).map_err(CommandError::from)
-}
-
-/// `list_unused_fields` の内部実装（テスト可能なように分離）。
-fn list_unused_fields_inner(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-) -> Result<Vec<UnusedFieldRow>, rusqlite::Error> {
-    // Step 1: 構造的参照（レイアウト・リレーション・バリューリスト）で used なフィールドを SQL で収集
-    let mut stmt = conn.prepare(
-        "WITH used_fields AS (
-           -- レイアウト上に直接配置されたフィールド
-           SELECT DISTINCT toc.base_table_name AS tbl, lfr.field_name
-           FROM layout_field_refs lfr
-           JOIN layouts l ON l.id = lfr.layout_id AND l.project_id = ?1
-           JOIN table_occurrences toc
-             ON toc.occurrence_name = lfr.table_occurrence AND toc.project_id = ?1
-           UNION
-           -- レイアウトオブジェクトのフィールド参照
-           SELECT DISTINCT toc.base_table_name AS tbl, lo.field_name
-           FROM layout_objects lo
-           JOIN layouts l ON l.id = lo.layout_id AND l.project_id = ?1
-           JOIN table_occurrences toc
-             ON toc.occurrence_name = lo.field_table_occurrence AND toc.project_id = ?1
-           WHERE lo.field_name IS NOT NULL AND lo.field_name != ''
-           UNION
-           -- リレーション結合キー（左側）
-           SELECT DISTINCT toc.base_table_name AS tbl, jp.left_field
-           FROM join_predicates jp
-           JOIN relationships r ON r.id = jp.relationship_id AND r.project_id = ?1
-           JOIN table_occurrences toc
-             ON toc.occurrence_name = r.left_table AND toc.project_id = ?1
-           WHERE jp.left_field != ''
-           UNION
-           -- リレーション結合キー（右側）
-           SELECT DISTINCT toc.base_table_name AS tbl, jp.right_field
-           FROM join_predicates jp
-           JOIN relationships r ON r.id = jp.relationship_id AND r.project_id = ?1
-           JOIN table_occurrences toc
-             ON toc.occurrence_name = r.right_table AND toc.project_id = ?1
-           WHERE jp.right_field != ''
-           UNION
-           -- バリューリストのフィールド参照（PrimaryField / SecondaryField）
-           SELECT DISTINCT toc.base_table_name AS tbl, vlfr.field_name
-           FROM value_list_field_refs vlfr
-           JOIN value_lists vl ON vl.id = vlfr.value_list_id AND vl.project_id = ?1
-           JOIN table_occurrences toc
-             ON toc.occurrence_name = vlfr.table_occurrence AND toc.project_id = ?1
-         )
-         SELECT bt.name, f.name, f.field_type, f.data_type, f.id
-         FROM fields f
-         JOIN base_tables bt ON bt.id = f.table_id AND bt.project_id = ?1
-         WHERE f.field_type NOT IN ('Summary')
-           AND NOT EXISTS (
-             SELECT 1 FROM used_fields uf
-             WHERE uf.tbl = bt.name AND uf.field_name = f.name
-           )
-         ORDER BY bt.name, f.name",
-    )?;
-    let mut rows = stmt
-        .query_map(params![project_id], |row| {
-            Ok(UnusedFieldRow {
-                table_name: row.get(0)?,
-                field_name: row.get(1)?,
-                field_type: row.get(2)?,
-                data_type: row.get(3)?,
-                field_id: row.get(4)?,
-            })
-        })?
-        .collect::<Result<Vec<_>, _>>()?;
-
-    // Step 2: スクリプト・フィールド計算式の OccName::FieldName Rust ポスト処理
-    // SQL の LIKE クロスジョインを避けるため Rust でフィルタリングする。
-    // 検出パターン: "OccurrenceName::FieldName" の文字列が計算テキストに含まれる場合に使用済みと判定。
-    if !rows.is_empty() {
-        // ベーステーブル名 → オカレンス名リスト のマップを構築
-        let occ_map = fetch_occ_map(conn, project_id)?;
-
-        // 全計算テキストを1つの文字列に結合（スクリプト + フィールド各 calc 列）
-        let all_calc_text = fetch_all_calc_texts(conn, project_id)?;
-
-        // 計算テキストに OccName::FieldName が含まれるフィールドを除外
-        rows.retain(|f| {
-            let occs = occ_map.get(&f.table_name).map(Vec::as_slice).unwrap_or(&[]);
-            !occs
-                .iter()
-                .any(|occ| all_calc_text.contains(&format!("{}::{}", occ, f.field_name)))
-        });
-    }
-
-    // Step 3: 同テーブルのフィールド計算式内ベア参照チェック（FileMaker 仕様準拠）
-    // FileMaker では計算式内のベア参照は常に同テーブルのフィールドを指す。
-    // スクリプトステップは実行時レイアウトコンテキスト依存のため対象外。
-    if !rows.is_empty() {
-        let same_table_calcs = fetch_same_table_field_calcs(conn, project_id)?;
-
-        rows.retain(|f| {
-            let calcs = same_table_calcs
-                .get(&f.table_name)
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            !calcs
-                .iter()
-                .any(|calc| has_bare_field_ref(calc, &f.field_name))
-        });
-    }
-
-    Ok(rows)
-}
-
-/// テーブル名 → 同テーブル内全フィールドの計算テキスト一覧を返す。
-///
-/// FileMaker では計算式内のベア参照は同テーブルのフィールドを指すため、
-/// 同テーブルのフィールド calc からのみベア参照チェックを行う。
-/// 対象列: calculation / auto_enter_calc / val_calc
-fn fetch_same_table_field_calcs(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-) -> Result<std::collections::HashMap<String, Vec<String>>, rusqlite::Error> {
-    let mut stmt = conn.prepare(
-        "SELECT bt.name,
-                COALESCE(f.calculation, ''),
-                COALESCE(f.auto_enter_calc, ''),
-                COALESCE(f.val_calc, '')
-         FROM fields f
-         JOIN base_tables bt ON bt.id = f.table_id AND bt.project_id = ?1",
-    )?;
-    let mut map: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
-    let rows = stmt.query_map(params![project_id], |row| {
-        Ok((
-            row.get::<_, String>(0)?,
-            row.get::<_, String>(1)?,
-            row.get::<_, String>(2)?,
-            row.get::<_, String>(3)?,
-        ))
-    })?;
-    for row in rows {
-        let (table, calc, ae, val) = row?;
-        let entry = map.entry(table).or_default();
-        if !calc.is_empty() {
-            entry.push(calc);
-        }
-        if !ae.is_empty() {
-            entry.push(ae);
-        }
-        if !val.is_empty() {
-            entry.push(val);
-        }
-    }
-    Ok(map)
-}
-
-/// ベーステーブル名 → オカレンス名リスト のマップを取得する。
-fn fetch_occ_map(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-) -> Result<std::collections::HashMap<String, Vec<String>>, rusqlite::Error> {
-    let mut stmt = conn.prepare(
-        "SELECT base_table_name, occurrence_name
-         FROM table_occurrences
-         WHERE project_id = ?1",
-    )?;
-    let mut map: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
-    let rows = stmt.query_map(params![project_id], |row| {
-        Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-    })?;
-    for row in rows {
-        let (base, occ) = row?;
-        map.entry(base).or_default().push(occ);
-    }
-    Ok(map)
-}
-
-/// プロジェクト内の全計算テキストを結合した文字列を返す。
-///
-/// 対象:
-/// - script_steps.calculation / step_text
-/// - fields.calculation / auto_enter_calc / val_calc
-fn fetch_all_calc_texts(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-) -> Result<String, rusqlite::Error> {
-    let mut parts: Vec<String> = Vec::new();
-
-    // スクリプトステップの計算式・ステップテキスト
-    {
-        let mut stmt = conn.prepare(
-            "SELECT COALESCE(ss.calculation, ''), COALESCE(ss.step_text, '')
-             FROM script_steps ss
-             JOIN scripts s ON s.id = ss.script_id AND s.project_id = ?1",
-        )?;
-        let rows = stmt.query_map(params![project_id], |row| {
-            Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-        })?;
-        for row in rows {
-            let (calc, text) = row?;
-            if !calc.is_empty() {
-                parts.push(calc);
-            }
-            if !text.is_empty() {
-                parts.push(text);
-            }
-        }
-    }
-
-    // フィールドの各計算式列（calculation / auto_enter_calc / val_calc）
-    {
-        let mut stmt = conn.prepare(
-            "SELECT COALESCE(f.calculation, ''),
-                    COALESCE(f.auto_enter_calc, ''),
-                    COALESCE(f.val_calc, '')
-             FROM fields f
-             JOIN base_tables bt ON bt.id = f.table_id AND bt.project_id = ?1",
-        )?;
-        let rows = stmt.query_map(params![project_id], |row| {
-            Ok((
-                row.get::<_, String>(0)?,
-                row.get::<_, String>(1)?,
-                row.get::<_, String>(2)?,
-            ))
-        })?;
-        for row in rows {
-            let (calc, ae, val) = row?;
-            if !calc.is_empty() {
-                parts.push(calc);
-            }
-            if !ae.is_empty() {
-                parts.push(ae);
-            }
-            if !val.is_empty() {
-                parts.push(val);
-            }
-        }
-    }
-
-    Ok(parts.join("\n"))
-}
-
-/// レイアウトフィールド参照のデバッグ情報を返す。
-#[tauri::command]
-pub async fn get_layout_ref_debug_info(
-    state: tauri::State<'_, AppState>,
-    project_id: i64,
-) -> Result<LayoutRefDebugInfo, CommandError> {
-    let db = super::lock_db(&state)?;
-    get_layout_ref_debug_info_inner(&db.conn, project_id).map_err(CommandError::from)
-}
-
-// ---------------------------------------------------------------------------
-// inner 関数（テスト可能・Tauri State 非依存）
-// ---------------------------------------------------------------------------
-
-/// `get_field_layout_refs` の内部実装。
-fn get_field_layout_refs_inner(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-    table_name: &str,
-    field_name: &str,
-) -> Result<Vec<FieldRefLayout>, rusqlite::Error> {
-    let mut stmt = conn.prepare(
-        "SELECT DISTINCT l.id, l.name
-         FROM layouts l
-         JOIN table_occurrences toc_main
-           ON toc_main.project_id = l.project_id
-          AND toc_main.occurrence_name = l.table_occurrence_name
-          AND toc_main.base_table_name = ?2
-         JOIN layout_field_refs lfr
-           ON lfr.layout_id = l.id
-          AND lfr.field_name = ?3
-         JOIN table_occurrences toc_field
-           ON toc_field.project_id = ?1
-          AND toc_field.occurrence_name = lfr.table_occurrence
-          AND toc_field.base_table_name = ?2
-         WHERE l.project_id = ?1
-         ORDER BY l.position, l.name",
-    )?;
-    let rows = stmt
-        .query_map(params![project_id, table_name, field_name], |row| {
-            Ok(FieldRefLayout {
-                layout_id: row.get(0)?,
-                layout_name: row.get(1)?,
-            })
-        })?
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(rows)
-}
-
-/// `resolve_layout_field` の内部実装。
-fn resolve_layout_field_inner(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-    occurrence_name: &str,
-    field_name: &str,
-) -> Result<Option<FieldLocation>, rusqlite::Error> {
-    conn.query_row(
-        "SELECT bt.id, f.id, bt.name
-           FROM fields f
-           JOIN base_tables bt ON bt.id = f.table_id
-           JOIN table_occurrences toc
-             ON toc.base_table_name = bt.name
-            AND toc.project_id = bt.project_id
-          WHERE bt.project_id = ?1
-            AND toc.occurrence_name = ?2
-            AND f.name = ?3
-          LIMIT 1",
-        params![project_id, occurrence_name, field_name],
-        |row| {
-            Ok(FieldLocation {
-                table_id: row.get(0)?,
-                field_id: row.get(1)?,
-                table_name: row.get(2)?,
-            })
-        },
-    )
-    .optional()
-}
-
-/// `get_field_relationship_keys` の内部実装。
-fn get_field_relationship_keys_inner(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-    table_name: &str,
-    field_name: &str,
-) -> Result<Vec<FieldRelKeyRef>, rusqlite::Error> {
-    let mut stmt = conn.prepare(
-        "SELECT r.id, r.name, r.left_table, r.right_table, jp.operator, 'left'
-         FROM relationships r
-         JOIN join_predicates jp ON jp.relationship_id = r.id
-         JOIN table_occurrences toc
-           ON toc.occurrence_name = r.left_table AND toc.project_id = r.project_id
-         WHERE r.project_id = ?1
-           AND jp.left_field = ?3
-           AND toc.base_table_name = ?2
-         UNION
-         SELECT r.id, r.name, r.left_table, r.right_table, jp.operator, 'right'
-         FROM relationships r
-         JOIN join_predicates jp ON jp.relationship_id = r.id
-         JOIN table_occurrences toc
-           ON toc.occurrence_name = r.right_table AND toc.project_id = r.project_id
-         WHERE r.project_id = ?1
-           AND jp.right_field = ?3
-           AND toc.base_table_name = ?2
-         ORDER BY r.name",
-    )?;
-    let rows = stmt
-        .query_map(params![project_id, table_name, field_name], |row| {
-            Ok(FieldRelKeyRef {
-                relationship_id: row.get(0)?,
-                relationship_name: row.get(1)?,
-                left_table: row.get(2)?,
-                right_table: row.get(3)?,
-                operator: row.get(4)?,
-                side: row.get(5)?,
-            })
-        })?
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(rows)
-}
-
-/// `get_layout_ref_debug_info` の内部実装。
-fn get_layout_ref_debug_info_inner(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-) -> Result<LayoutRefDebugInfo, rusqlite::Error> {
-    let occurrence_count: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM table_occurrences WHERE project_id = ?1",
-            params![project_id],
-            |r| r.get(0),
-        )
-        .unwrap_or(0);
-
-    let layout_field_ref_count: i64 = conn
-        .query_row(
-            "SELECT COUNT(*) FROM layout_field_refs lfr
-             JOIN layouts l ON l.id = lfr.layout_id
-             WHERE l.project_id = ?1",
-            params![project_id],
-            |r| r.get(0),
-        )
-        .unwrap_or(0);
-
-    let mut stmt = conn.prepare(
-        "SELECT occurrence_name || ' -> ' || base_table_name
-         FROM table_occurrences WHERE project_id = ?1 LIMIT 10",
-    )?;
-    let sample_occurrences = stmt
-        .query_map(params![project_id], |r| r.get::<_, String>(0))?
-        .filter_map(|r| r.ok())
-        .collect();
-
-    let mut stmt2 = conn.prepare(
-        "SELECT l.name || ' | ' || lfr.table_occurrence || '::' || lfr.field_name
-         FROM layout_field_refs lfr
-         JOIN layouts l ON l.id = lfr.layout_id
-         WHERE l.project_id = ?1 LIMIT 10",
-    )?;
-    let sample_field_refs = stmt2
-        .query_map(params![project_id], |r| r.get::<_, String>(0))?
-        .filter_map(|r| r.ok())
-        .collect();
-
-    Ok(LayoutRefDebugInfo {
-        occurrence_count,
-        layout_field_ref_count,
-        sample_occurrences,
-        sample_field_refs,
-    })
-}
-
-// ---------------------------------------------------------------------------
-// 内部ヘルパー
-// ---------------------------------------------------------------------------
-
-/// ベーステーブルに紐づく全オカレンス名を返す。
-fn fetch_occ_names(
-    conn: &rusqlite::Connection,
-    project_id: i64,
-    base_table_name: &str,
-) -> Result<Vec<String>, rusqlite::Error> {
-    let mut stmt = conn.prepare(
-        "SELECT occurrence_name FROM table_occurrences
-         WHERE project_id = ?1 AND base_table_name = ?2",
-    )?;
-    let rows = stmt
-        .query_map(params![project_id, base_table_name], |r| r.get(0))?
-        .collect::<Result<Vec<String>, _>>()?;
-    Ok(rows)
-}
-
-/// FileMaker 計算式・スクリプトテキスト中でフィールドが参照されているか判定。
-///
-/// - `OccName::field_name` 形式（任意のオカレンス名）
-/// - 識別子境界を考慮した bare `field_name` 形式
-///
-/// のいずれかにマッチすれば `true` を返す。
-fn field_ref_matches(text: &str, occ_names: &[String], field_name: &str) -> bool {
-    if text.is_empty() {
-        return false;
-    }
-    // OccName::FieldName パターン
-    if occ_names
-        .iter()
-        .any(|occ| text.contains(&format!("{}::{}", occ, field_name)))
-    {
-        return true;
-    }
-    // bare FieldName パターン（識別子境界チェック）
-    has_bare_field_ref(text, field_name)
-}
-
-/// `text` 中に `field_name` が識別子として単独で現れるか判定。
-///
-/// 直前が識別子文字または `:` でなく、直後が識別子文字または `(` でない位置に
-/// `field_name` が存在する場合に `true` を返す。
-fn has_bare_field_ref(text: &str, field_name: &str) -> bool {
-    let mut from = 0;
-    while let Some(pos) = text[from..].find(field_name) {
-        let abs = from + pos;
-        let end = abs + field_name.len();
-        let before_ok = text[..abs]
-            .chars()
-            .last()
-            .is_none_or(|c| !is_fm_ident_char(c) && c != ':');
-        let after_ok = text[end..]
-            .chars()
-            .next()
-            .is_none_or(|c| !is_fm_ident_char(c) && c != '(');
-        if before_ok && after_ok {
-            return true;
-        }
-        // field_name.len() バイト分進める（UTF-8 境界が保証される）
-        from = abs + field_name.len();
-    }
-    false
-}
-
-/// FileMaker 識別子文字（フィールド名・TO名に使える文字）の判定。
-fn is_fm_ident_char(c: char) -> bool {
-    c.is_alphanumeric() || c == '_' || c == '＿'
-}
-
-// ---------------------------------------------------------------------------
-// テスト
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rusqlite::Connection;
-
-    use crate::db::schema::initialize;
-
-    // -----------------------------------------------------------------------
-    // ヘルパー: has_bare_field_ref / field_ref_matches の単体テスト
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn bare_ref_detects_standalone() {
-        // フィールド名が単独で現れる
-        assert!(has_bare_field_ref("Amount * 1.1", "Amount"));
-        assert!(has_bare_field_ref("Amount", "Amount"));
-        assert!(has_bare_field_ref("(Amount + Tax)", "Amount"));
-    }
-
-    #[test]
-    fn bare_ref_ignores_qualified() {
-        // OccName::FieldName は bare ref ではない（`:` が直前にある）
-        assert!(!has_bare_field_ref("Invoice::Amount * 1.1", "Amount"));
-        assert!(!has_bare_field_ref("Inv::Amount", "Amount"));
-    }
-
-    #[test]
-    fn bare_ref_ignores_substring() {
-        // 別フィールド名の一部にマッチしない
-        assert!(!has_bare_field_ref("TotalAmount + 1", "Amount"));
-        assert!(!has_bare_field_ref("合計金額 + 1", "金額"));
-    }
-
-    #[test]
-    fn bare_ref_ignores_function_call() {
-        // 関数呼び出し（直後が `(`）は除外
-        assert!(!has_bare_field_ref("Amount(x)", "Amount"));
-    }
-
-    #[test]
-    fn field_ref_matches_via_occ_name() {
-        // OccName が base_table_name と異なる場合も検出できる
-        let occ_names = vec!["InvoiceAlias".to_string()];
-        assert!(field_ref_matches(
-            "InvoiceAlias::Amount * 1.1",
-            &occ_names,
-            "Amount"
-        ));
-        // base_table_name（Invoice）では検出されない
-        let empty: Vec<String> = vec![];
-        assert!(!field_ref_matches("InvoiceAlias::Amount", &empty, "Amount"));
-    }
-
-    // -----------------------------------------------------------------------
-    // ヘルパー: DB セットアップ
-    // -----------------------------------------------------------------------
-
-    /// インメモリ DB を作り、プロジェクト・テーブル・オカレンス・フィールドを挿入する。
-    ///
-    /// テーブル構成:
-    /// - Invoice (occurrence: "Invoice", "InvoiceAlias")
-    /// - Order   (occurrence: "Order")
-    ///
-    /// フィールド:
-    /// - Invoice::Amount          計算式なし
-    /// - Invoice::Total           計算式 "InvoiceAlias::Amount * 1.1"  (オカレンス名経由)
-    /// - Invoice::合計金額         計算式 "Amount + Tax"                (bare ref + 部分一致の罠)
-    /// - Order::Total             計算式 "Invoice::Amount + 100"        (base table 名と一致)
-    /// - Order::Note              計算式 "Order::Amount"                (別フィールド)
-    fn setup() -> (Connection, i64) {
-        let conn = Connection::open_in_memory().unwrap();
-        initialize(&conn).unwrap();
-
-        conn.execute("INSERT INTO solutions(name) VALUES('sol')", [])
-            .unwrap();
-        let solution_id = conn.last_insert_rowid();
-        conn.execute(
-            "INSERT INTO projects(solution_id, name, fm_version) VALUES(?1, 'test.fmp12', '19')",
-            [solution_id],
-        )
-        .unwrap();
-        let project_id = conn.last_insert_rowid();
-
-        // base_tables
-        conn.execute(
-            "INSERT INTO base_tables(project_id, fm_id, name) VALUES(?1, 1, 'Invoice')",
-            [project_id],
-        )
-        .unwrap();
-        let invoice_table_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "INSERT INTO base_tables(project_id, fm_id, name) VALUES(?1, 2, 'Order')",
-            [project_id],
-        )
-        .unwrap();
-        let order_table_id = conn.last_insert_rowid();
-
-        // table_occurrences（Invoice に 2 つのオカレンス）
-        conn.execute(
-            "INSERT INTO table_occurrences(project_id, occurrence_name, base_table_name)
-             VALUES(?1, 'Invoice', 'Invoice')",
-            [project_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO table_occurrences(project_id, occurrence_name, base_table_name)
-             VALUES(?1, 'InvoiceAlias', 'Invoice')",
-            [project_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO table_occurrences(project_id, occurrence_name, base_table_name)
-             VALUES(?1, 'Order', 'Order')",
-            [project_id],
-        )
-        .unwrap();
-
-        // fields
-        for (fm_id, table_id, name, calc) in [
-            (1, invoice_table_id, "Amount", ""),
-            (2, invoice_table_id, "Total", "InvoiceAlias::Amount * 1.1"),
-            (3, invoice_table_id, "合計金額", "Amount + Tax"), // bare ref to Amount
-            (4, order_table_id, "Total", "Invoice::Amount + 100"),
-            (5, order_table_id, "Note", "Order::Amount"),
-        ] {
-            conn.execute(
-                "INSERT INTO fields(project_id, table_id, fm_id, name, field_type, data_type, calculation)
-                 VALUES(?1, ?2, ?3, ?4, 'Calculation', 'Number', ?5)",
-                rusqlite::params![project_id, table_id, fm_id, name, calc],
-            )
-            .unwrap();
-        }
-
-        (conn, project_id)
-    }
-
-    /// `get_field_calc_refs` の内部ロジックを直接呼び出すヘルパー。
-    fn calc_refs(
-        conn: &Connection,
-        project_id: i64,
-        table_name: &str,
-        field_name: &str,
-    ) -> Vec<(String, String)> {
-        let occ_names = fetch_occ_names(conn, project_id, table_name).unwrap();
-        let mut stmt = conn
-            .prepare(
-                "SELECT f.id, f.name, bt.name, bt.id, f.calculation
-                 FROM fields f
-                 JOIN base_tables bt ON bt.id = f.table_id
-                 WHERE bt.project_id = ?1
-                   AND f.calculation IS NOT NULL AND f.calculation != ''
-                   AND NOT (bt.name = ?2 AND f.name = ?3)
-                 ORDER BY bt.name, f.name",
-            )
-            .unwrap();
-        stmt.query_map(
-            rusqlite::params![project_id, table_name, field_name],
-            |row| {
-                Ok((
-                    row.get::<_, i64>(0)?,
-                    row.get::<_, String>(1)?,
-                    row.get::<_, String>(2)?,
-                    row.get::<_, i64>(3)?,
-                    row.get::<_, String>(4)?,
-                ))
-            },
-        )
-        .unwrap()
-        .filter_map(|r| r.ok())
-        .filter(|(_, _, _, _, calc)| field_ref_matches(calc, &occ_names, field_name))
-        .map(|(_, fname, tname, _, _)| (tname, fname))
-        .collect()
-    }
-
-    // -----------------------------------------------------------------------
-    // get_field_calc_refs 相当のテスト
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn test_finds_via_occurrence_name() {
-        let (conn, project_id) = setup();
-        let refs = calc_refs(&conn, project_id, "Invoice", "Amount");
-        // Invoice::Total (InvoiceAlias::Amount), Invoice::合計金額 (bare Amount), Order::Total (Invoice::Amount)
-        assert!(refs.contains(&("Invoice".to_string(), "Total".to_string())));
-        assert!(refs.contains(&("Invoice".to_string(), "合計金額".to_string())));
-        assert!(refs.contains(&("Order".to_string(), "Total".to_string())));
-    }
-
-    #[test]
-    fn test_excludes_self() {
-        let (conn, project_id) = setup();
-        let refs = calc_refs(&conn, project_id, "Invoice", "Amount");
-        assert!(!refs.contains(&("Invoice".to_string(), "Amount".to_string())));
-    }
-
-    #[test]
-    fn test_returns_empty_when_no_refs() {
-        let (conn, project_id) = setup();
-        // Order::Note は Invoice::Amount を参照していない
-        let refs = calc_refs(&conn, project_id, "Invoice", "Note");
-        assert!(refs.is_empty());
-    }
-
-    #[test]
-    fn test_different_occurrence_not_matched() {
-        let (conn, project_id) = setup();
-        // Order::Note は Order::Amount を参照しており Invoice::Amount ではない
-        let refs = calc_refs(&conn, project_id, "Invoice", "Amount");
-        assert!(!refs.contains(&("Order".to_string(), "Note".to_string())));
-    }
-
-    // -----------------------------------------------------------------------
-    // list_unused_fields 相当のヘルパー
-    // -----------------------------------------------------------------------
-
-    /// `list_unused_fields_inner` を呼び出して (table_name, field_name) のペアを返す。
-    fn list_unused(conn: &Connection, project_id: i64) -> Vec<(String, String)> {
-        list_unused_fields_inner(conn, project_id)
-            .unwrap()
-            .into_iter()
-            .map(|r| (r.table_name, r.field_name))
-            .collect()
-    }
-
-    // -----------------------------------------------------------------------
-    // list_unused_fields: 計算式参照のテスト
-    // -----------------------------------------------------------------------
-
-    /// setup() の Invoice::Amount は "InvoiceAlias::Amount" として Invoice::Total の計算式で
-    /// 参照されている。新仕様では未参照リストに含まれてはいけない。
-    #[test]
-    fn unused_fields_excludes_calc_referenced_via_occurrence() {
-        let (conn, project_id) = setup();
-        // script_steps に計算式を挿入
-        conn.execute(
-            "INSERT INTO scripts(project_id, fm_id, name) VALUES(?1, 1, 'TestScript')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let script_id = conn.last_insert_rowid();
-        conn.execute(
-            "INSERT INTO script_steps(script_id, step_type_id, name, enabled, calculation, position)
-             VALUES(?1, 89, 'Set Field', 1, 'InvoiceAlias::Amount * 1.1', 0)",
-            rusqlite::params![script_id],
-        )
-        .unwrap();
-
-        let unused = list_unused(&conn, project_id);
-        // Invoice::Amount は InvoiceAlias::Amount として script_steps に参照されているので
-        // 未参照リストに出てはいけない
-        assert!(
-            !unused.iter().any(|(t, f)| t == "Invoice" && f == "Amount"),
-            "Invoice::Amount は計算式で参照されているのに未参照リストに含まれた: {:?}",
-            unused
-        );
-    }
-
-    /// fields.calculation で参照されているフィールドも未参照リストから除外される。
-    #[test]
-    fn unused_fields_excludes_field_calc_referenced() {
-        let (conn, project_id) = setup();
-        // setup() では Invoice::Total の calculation = "InvoiceAlias::Amount * 1.1"
-        // Invoice::Amount は既に fields テーブルに OccName::FieldName 形式で参照されている
-
-        let unused = list_unused(&conn, project_id);
-        // Invoice::Amount は Invoice::Total の計算式で参照されているので未参照リストに出ない
-        assert!(
-            !unused.iter().any(|(t, f)| t == "Invoice" && f == "Amount"),
-            "Invoice::Amount はフィールド計算式で参照されているのに未参照リストに含まれた: {:?}",
-            unused
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // list_unused_fields: バリューリスト参照のテスト
-    // -----------------------------------------------------------------------
-
-    fn setup_with_value_list_ref() -> (Connection, i64) {
-        let (conn, project_id) = setup();
-
-        // Invoice::Amount を参照するバリューリストを追加
-        conn.execute(
-            "INSERT INTO value_lists(project_id, fm_id, name, source) VALUES(?1, 1, 'TestVL', 'Field')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let vl_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "INSERT INTO value_list_field_refs(value_list_id, table_occurrence, field_name)
-             VALUES(?1, 'Invoice', 'Amount')",
-            rusqlite::params![vl_id],
-        )
-        .unwrap();
-
-        (conn, project_id)
-    }
-
-    /// バリューリストのフィールド参照として使われているフィールドは未参照リストに出ない。
-    #[test]
-    fn unused_fields_excludes_value_list_field_refs() {
-        let (conn, project_id) = setup_with_value_list_ref();
-        let unused = list_unused(&conn, project_id);
-        assert!(
-            !unused.iter().any(|(t, f)| t == "Invoice" && f == "Amount"),
-            "Invoice::Amount はバリューリストで参照されているのに未参照リストに含まれた: {:?}",
-            unused
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // list_unused_fields: 同テーブル内ベア参照のテスト
-    // -----------------------------------------------------------------------
-
-    fn setup_with_bare_ref() -> (Connection, i64) {
-        let (conn, project_id) = setup();
-
-        // Customer テーブルを追加
-        conn.execute(
-            "INSERT INTO base_tables(project_id, fm_id, name) VALUES(?1, 10, 'Customer')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let customer_table_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "INSERT INTO table_occurrences(project_id, occurrence_name, base_table_name)
-             VALUES(?1, 'Customer', 'Customer')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-
-        // Status フィールド: レイアウト・リレーション・OccName::FieldName いずれでも未参照
-        conn.execute(
-            "INSERT INTO fields(project_id, table_id, fm_id, name, field_type, data_type, calculation)
-             VALUES(?1, ?2, 101, 'Status', 'Normal', 'Text', '')",
-            rusqlite::params![project_id, customer_table_id],
-        )
-        .unwrap();
-
-        // DisplayName フィールド: Status をベア参照する計算式
-        conn.execute(
-            "INSERT INTO fields(project_id, table_id, fm_id, name, field_type, data_type, calculation)
-             VALUES(?1, ?2, 102, 'DisplayName', 'Calculation', 'Text', 'Upper(Status)')",
-            rusqlite::params![project_id, customer_table_id],
-        )
-        .unwrap();
-
-        (conn, project_id)
-    }
-
-    /// 同テーブルの計算フィールドがベア参照しているフィールドは未参照リストに出ない。
-    #[test]
-    fn unused_fields_excludes_bare_ref_in_same_table_calc() {
-        let (conn, project_id) = setup_with_bare_ref();
-        let unused = list_unused(&conn, project_id);
-        // Customer::Status は Customer::DisplayName の計算式 "Upper(Status)" でベア参照されている
-        // → 未参照リストに含まれてはいけない
-        assert!(
-            !unused
-                .iter()
-                .any(|(t, f)| t == "Customer" && f == "Status"),
-            "Customer::Status は同テーブル calc でベア参照されているのに未参照リストに含まれた: {:?}",
-            unused
-        );
-    }
-
-    /// 異なるテーブルの同名フィールドはベア参照で誤検出されない。
-    #[test]
-    fn unused_fields_bare_ref_does_not_cross_table() {
-        let (conn, project_id) = setup_with_bare_ref();
-        let unused = list_unused(&conn, project_id);
-        // Order テーブルには "Status" フィールドが存在しないが、
-        // 別テーブルの calc が "Status" を使っていても Order には影響しない。
-        // setup() の Invoice テーブルには "Status" フィールドがないので存在確認のみ。
-        // Customer::DisplayName は Customer テーブルのフィールドなので
-        // Invoice テーブルのフィールドには影響しないことを確認。
-        // Invoice::Amount は Customer の calc でベア参照されていないので未参照のまま。
-        // （Invoice::Amount は setup() の Invoice::Total が OccName 参照しているので実際は除外済み）
-        // ここでは Invoice テーブルのフィールドが Customer の bare ref で誤って除外されないことを確認。
-        let customer_bare_referenced: Vec<_> =
-            unused.iter().filter(|(t, _)| t == "Invoice").collect();
-        // Invoice テーブルのフィールドが Customer::DisplayName の "Upper(Status)" によって
-        // 除外されていないことを確認（Invoice に Status というフィールドはないので変化なし）
-        let _ = customer_bare_referenced; // setup 内容次第なので存在確認のみ
-    }
-
-    // -----------------------------------------------------------------------
-    // ヘルパー関数の直接テスト
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn fetch_occ_names_returns_names_for_table() {
-        let (conn, project_id) = setup();
-        let names = fetch_occ_names(&conn, project_id, "Invoice").unwrap();
-        assert!(names.contains(&"Invoice".to_string()));
-        assert!(names.contains(&"InvoiceAlias".to_string()));
-        assert_eq!(names.len(), 2);
-    }
-
-    #[test]
-    fn fetch_occ_names_returns_empty_for_unknown_table() {
-        let (conn, project_id) = setup();
-        let names = fetch_occ_names(&conn, project_id, "NonExistent").unwrap();
-        assert!(names.is_empty());
-    }
-
-    #[test]
-    fn fetch_occ_map_groups_by_base_table() {
-        let (conn, project_id) = setup();
-        let map = fetch_occ_map(&conn, project_id).unwrap();
-        let invoice_occs = map.get("Invoice").unwrap();
-        assert!(invoice_occs.contains(&"Invoice".to_string()));
-        assert!(invoice_occs.contains(&"InvoiceAlias".to_string()));
-        let order_occs = map.get("Order").unwrap();
-        assert!(order_occs.contains(&"Order".to_string()));
-    }
-
-    #[test]
-    fn fetch_all_calc_texts_includes_field_calcs() {
-        let (conn, project_id) = setup();
-        let text = fetch_all_calc_texts(&conn, project_id).unwrap();
-        // Invoice::Total の計算式が含まれる
-        assert!(text.contains("InvoiceAlias::Amount"));
-    }
-
-    #[test]
-    fn fetch_all_calc_texts_includes_script_step_calcs() {
-        let (conn, project_id) = setup();
-        // スクリプトステップを追加
-        conn.execute(
-            "INSERT INTO scripts(project_id, fm_id, name) VALUES(?1, 99, 'S')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let sid = conn.last_insert_rowid();
-        conn.execute(
-            "INSERT INTO script_steps(script_id, step_type_id, name, enabled, calculation, position)
-             VALUES(?1, 89, 'Set Field', 1, 'Invoice::Amount * 2', 0)",
-            rusqlite::params![sid],
-        )
-        .unwrap();
-        let text = fetch_all_calc_texts(&conn, project_id).unwrap();
-        assert!(text.contains("Invoice::Amount * 2"));
-    }
-
-    #[test]
-    fn fetch_same_table_field_calcs_groups_by_table() {
-        let (conn, project_id) = setup();
-        let map = fetch_same_table_field_calcs(&conn, project_id).unwrap();
-        let inv_calcs = map.get("Invoice").unwrap();
-        // Invoice::Total の calculation が含まれる
-        assert!(inv_calcs.iter().any(|c| c.contains("InvoiceAlias::Amount")));
-    }
-
-    // -----------------------------------------------------------------------
-    // get_field_layout_refs_inner のテスト
-    // -----------------------------------------------------------------------
-
-    fn setup_with_layout_refs() -> (Connection, i64) {
-        let (conn, project_id) = setup();
-
-        // レイアウト: Invoice をメインテーブルとして使用
-        conn.execute(
-            "INSERT INTO layouts(project_id, fm_id, name, table_occurrence_name, position)
-             VALUES(?1, 1, 'InvoiceList', 'Invoice', 0)",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let layout_id = conn.last_insert_rowid();
-
-        // layout_field_refs: Invoice::Amount を配置
-        conn.execute(
-            "INSERT INTO layout_field_refs(layout_id, table_occurrence, field_name)
-             VALUES(?1, 'Invoice', 'Amount')",
-            rusqlite::params![layout_id],
-        )
-        .unwrap();
-
-        // layout_field_refs: InvoiceAlias::Total も配置（別オカレンス名経由）
-        conn.execute(
-            "INSERT INTO layout_field_refs(layout_id, table_occurrence, field_name)
-             VALUES(?1, 'InvoiceAlias', 'Total')",
-            rusqlite::params![layout_id],
-        )
-        .unwrap();
-
-        (conn, project_id)
-    }
-
-    #[test]
-    fn layout_refs_returns_layout_containing_field() {
-        let (conn, project_id) = setup_with_layout_refs();
-        let refs = get_field_layout_refs_inner(&conn, project_id, "Invoice", "Amount").unwrap();
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].layout_name, "InvoiceList");
-    }
-
-    #[test]
-    fn layout_refs_returns_layout_via_alias_occurrence() {
-        let (conn, project_id) = setup_with_layout_refs();
-        // InvoiceAlias 経由で配置された Total フィールド
-        let refs = get_field_layout_refs_inner(&conn, project_id, "Invoice", "Total").unwrap();
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].layout_name, "InvoiceList");
-    }
-
-    #[test]
-    fn layout_refs_returns_empty_for_different_table_field() {
-        let (conn, project_id) = setup_with_layout_refs();
-        // Order テーブルのフィールドはこのレイアウトに配置されていない
-        let refs = get_field_layout_refs_inner(&conn, project_id, "Order", "Total").unwrap();
-        assert!(refs.is_empty());
-    }
-
-    // -----------------------------------------------------------------------
-    // resolve_layout_field_inner のテスト
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn resolve_layout_field_returns_location_for_existing_field() {
-        let (conn, project_id) = setup();
-        let loc = resolve_layout_field_inner(&conn, project_id, "Invoice", "Amount").unwrap();
-        assert!(loc.is_some());
-        let loc = loc.unwrap();
-        assert_eq!(loc.table_name, "Invoice");
-    }
-
-    #[test]
-    fn resolve_layout_field_returns_location_via_alias() {
-        let (conn, project_id) = setup();
-        // InvoiceAlias も Invoice ベーステーブルを指すので解決できる
-        let loc = resolve_layout_field_inner(&conn, project_id, "InvoiceAlias", "Amount").unwrap();
-        assert!(loc.is_some());
-        let loc = loc.unwrap();
-        assert_eq!(loc.table_name, "Invoice");
-    }
-
-    #[test]
-    fn resolve_layout_field_returns_none_for_unknown_field() {
-        let (conn, project_id) = setup();
-        let loc = resolve_layout_field_inner(&conn, project_id, "Invoice", "NonExistent").unwrap();
-        assert!(loc.is_none());
-    }
-
-    #[test]
-    fn resolve_layout_field_returns_none_for_unknown_occurrence() {
-        let (conn, project_id) = setup();
-        let loc = resolve_layout_field_inner(&conn, project_id, "UnknownOcc", "Amount").unwrap();
-        assert!(loc.is_none());
-    }
-
-    // -----------------------------------------------------------------------
-    // get_field_relationship_keys_inner のテスト
-    // -----------------------------------------------------------------------
-
-    fn setup_with_relationships() -> (Connection, i64) {
-        let (conn, project_id) = setup();
-
-        // リレーション: Invoice::Amount = Order::Total
-        conn.execute(
-            "INSERT INTO relationships(project_id, fm_id, name, left_table, right_table)
-             VALUES(?1, 1, 'Inv_Order', 'Invoice', 'Order')",
-            rusqlite::params![project_id],
-        )
-        .unwrap();
-        let rel_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "INSERT INTO join_predicates(relationship_id, left_field, right_field, operator, position)
-             VALUES(?1, 'Amount', 'Total', '=', 0)",
-            rusqlite::params![rel_id],
-        )
-        .unwrap();
-
-        (conn, project_id)
-    }
-
-    #[test]
-    fn rel_keys_detects_left_side_key() {
-        let (conn, project_id) = setup_with_relationships();
-        let refs =
-            get_field_relationship_keys_inner(&conn, project_id, "Invoice", "Amount").unwrap();
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].relationship_name, "Inv_Order");
-        assert_eq!(refs[0].side, "left");
-    }
-
-    #[test]
-    fn rel_keys_detects_right_side_key() {
-        let (conn, project_id) = setup_with_relationships();
-        let refs = get_field_relationship_keys_inner(&conn, project_id, "Order", "Total").unwrap();
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].relationship_name, "Inv_Order");
-        assert_eq!(refs[0].side, "right");
-    }
-
-    #[test]
-    fn rel_keys_returns_empty_when_no_match() {
-        let (conn, project_id) = setup_with_relationships();
-        let refs = get_field_relationship_keys_inner(&conn, project_id, "Invoice", "Note").unwrap();
-        assert!(refs.is_empty());
-    }
-
-    // -----------------------------------------------------------------------
-    // get_layout_ref_debug_info_inner のテスト
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn debug_info_counts_occurrences_and_refs() {
-        let (conn, project_id) = setup_with_layout_refs();
-        let info = get_layout_ref_debug_info_inner(&conn, project_id).unwrap();
-        // setup() で 3 オカレンス（Invoice, InvoiceAlias, Order）
-        assert_eq!(info.occurrence_count, 3);
-        // setup_with_layout_refs() で 2 件の layout_field_refs
-        assert_eq!(info.layout_field_ref_count, 2);
-        assert!(!info.sample_occurrences.is_empty());
-        assert!(!info.sample_field_refs.is_empty());
-    }
-
-    #[test]
-    fn debug_info_returns_zero_for_empty_project() {
-        let conn = Connection::open_in_memory().unwrap();
-        crate::db::schema::initialize(&conn).unwrap();
-        conn.execute("INSERT INTO solutions(name) VALUES('s')", [])
-            .unwrap();
-        let sid = conn.last_insert_rowid();
-        conn.execute(
-            "INSERT INTO projects(solution_id, name, fm_version) VALUES(?1, 'p', '19')",
-            [sid],
-        )
-        .unwrap();
-        let project_id = conn.last_insert_rowid();
-
-        let info = get_layout_ref_debug_info_inner(&conn, project_id).unwrap();
-        assert_eq!(info.occurrence_count, 0);
-        assert_eq!(info.layout_field_ref_count, 0);
-        assert!(info.sample_occurrences.is_empty());
-        assert!(info.sample_field_refs.is_empty());
-    }
-}
